Silence some more warnings INT2PTR.
p4raw-id: //depot/perl@11972

+++ taint.c
@@ -15,7 +15,7 @@ Perl_taint_proper(pTHX_ const char *f, const char *s)
 
 #ifdef HAS_SETEUID
     DEBUG_u(PerlIO_printf(Perl_debug_log,
-            "%s %d %"Uid_t_f" %"Uid_t_f"\n", s, PL_tainted, PL_uid, PL_euid));
+            "%s %d %"Uid_t_f" %"Uid_t_f"\n", s, PL_tainted, (long unsigned int)PL_uid, (long unsigned int)PL_euid));
 #endif
 
     if (PL_tainted) {
